Made HttpMethodConfiguration more pure bean like by moving utility methods into a separate class
