+++ wagon-providers/wagon-http/src/main/java/org/apache/maven/wagon/providers/http/AbstractHttpClientWagon.java
@@ -702,7 +702,7 @@ public abstract class AbstractHttpClientWagon
 
         if ( config != null )
         {
-            config.applyConfig(requestConfigBuilder);
+            ConfigurationUtils.copyConfig(config, requestConfigBuilder );
 
             if ( config.isUsePreemptive() && authenticationInfo != null )
             {
